+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 Adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) and
 [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.16.2] - 2021-11-06
+
+### Added
+
+- Can concatenate with `.of`
+- `to=` arg to `pretty_print`
+- `Utils.choose_table_format`
+
 ## [0.16.2] - 2021-11-05
 
 ### Added